rename hll.h/cc hllc.h/cc in prep to create an HLL class, which by convention will use the hll file name
